+++ bk_JUCE/bitKlavier/Source/SynchronicViewController.cpp
@@ -1505,7 +1505,7 @@ void SynchronicModificationEditor::BKADSRSliderValueChanged(String name, int att
     {
         if(envelopeSliders[i]->getName() == name)
         {
-            envs.set(i, {attack, decay, sustain, release, 1, i});
+            envs.set(i, {(float)attack, decay, sustain, release, 1, i});
             envelopeSliders[i]->setActive();
             envelopeSliders[i]->setBright();
         }
